Fold ConvertPoclExit pass into UnreachablesToReturns as a helper function
Reduces the number of separate passes in the DeSPMD pipeline by moving
the __pocl_exit/__pocl_trap conversion logic into a static helper called
at the start of UnreachablesToReturns::run().
