+++ lib/llvmopencl/CMakeLists.txt
@@ -37,8 +37,6 @@ set(LLVMPASSES_SOURCES "AllocasToEntry.cc"
                        "BarrierTailReplication.h"
                        "CanonicalizeBarriers.cc"
                        "CanonicalizeBarriers.h"
-                       "ConvertPoclExit.cc"
-                       "ConvertPoclExit.h"
                        "DebugHelpers.cc"
                        "DebugHelpers.h"
                        "EmitPrintf.cc"
